+++ src/main/java/KUSITMS/WITHUS/domain/organization/entity/Organization.java
@@ -1,7 +1,7 @@
 package KUSITMS.WITHUS.domain.organization.entity;
 
 import KUSITMS.WITHUS.domain.application.position.entity.Position;
-import KUSITMS.WITHUS.domain.user.entity.UserOrganization;
+import KUSITMS.WITHUS.domain.user.userOrganization.entity.UserOrganization;
 import KUSITMS.WITHUS.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
