+++ src/main/java/KUSITMS/WITHUS/domain/user/user/entity/User.java
@@ -1,8 +1,9 @@
-package KUSITMS.WITHUS.domain.user.entity;
+package KUSITMS.WITHUS.domain.user.user.entity;
 
 import KUSITMS.WITHUS.domain.application.entity.Comment;
 import KUSITMS.WITHUS.domain.interview.timeslotUser.entity.TimeSlotUser;
-import KUSITMS.WITHUS.domain.user.enumerate.Role;
+import KUSITMS.WITHUS.domain.user.userOrganization.entity.UserOrganization;
+import KUSITMS.WITHUS.domain.user.user.enumerate.Role;
 import KUSITMS.WITHUS.global.common.BaseEntity;
 import KUSITMS.WITHUS.global.common.enumerate.Gender;
 import jakarta.persistence.*;
