+++ src/main/java/KUSITMS/WITHUS/domain/user/user/service/UserService.java
@@ -1,0 +1,9 @@
+package KUSITMS.WITHUS.domain.user.user.service;
+
+import KUSITMS.WITHUS.domain.user.user.dto.UserRequestDTO;
+import KUSITMS.WITHUS.domain.user.user.entity.User;
+
+public interface UserService {
+    User getUserByEmail(String email);
+    void joinProcess(UserRequestDTO.Join request);
+}
